+++ sources/org/mycore/backend/xmldb/MCRXMLDBEventHandlerIndexMeta.java
@@ -97,7 +97,7 @@ public class MCRXMLDBEventHandlerIndexMeta extends MCREventHandlerBase {
 
         // save the stop time
         long t2 = System.currentTimeMillis();
-        double diff = (double) (t2 - t1) / 1000.0;
+        double diff = (t2 - t1) / 1000.0;
         LOGGER.debug("MCRXMLDBEventHandlerIndexMeta create: done in " + diff + " sec.");
     }
 
@@ -148,7 +148,7 @@ public class MCRXMLDBEventHandlerIndexMeta extends MCREventHandlerBase {
 
         // save the stop time
         long t2 = System.currentTimeMillis();
-        double diff = (double) (t2 - t1) / 1000.0;
+        double diff = (t2 - t1) / 1000.0;
         LOGGER.debug("MCRXMLDBEventHandlerIndexMeta update: done in " + diff + " sec.");
     }
 
@@ -186,7 +186,7 @@ public class MCRXMLDBEventHandlerIndexMeta extends MCREventHandlerBase {
 
         // save the stop time
         long t2 = System.currentTimeMillis();
-        double diff = (double) (t2 - t1) / 1000.0;
+        double diff = (t2 - t1) / 1000.0;
         LOGGER.debug("MCRXMLDBEventHandlerIndexMeta delete: done in " + diff + " sec.");
     }
 
@@ -237,7 +237,7 @@ public class MCRXMLDBEventHandlerIndexMeta extends MCREventHandlerBase {
 
         // save the stop time
         long t2 = System.currentTimeMillis();
-        double diff = (double) (t2 - t1) / 1000.0;
+        double diff = (t2 - t1) / 1000.0;
         LOGGER.debug("MCRXMLDBEventHandlerIndexMeta repair: done in " + diff + " sec.");
     }
